Implement the simplest possible task execution
Resolves: #2

This commit implements task object detection and argo workflow
creation, which combined faciliate a basic form of task execution as
specified in #2.
The task and workflow are linked by using the same name, this
should be revisited in the future and was left open in #2.
More complex aspects such as idempotency or task execution failures are
left for future commits, as they are not required for #2.

Extra:
A few changes were made to the test setup and tear down to start the
controller in the test environment.