Implement minimum viable Task specification
Inorder to progress the development of AME a minimum specification is
required which contains the bare minimum for AME to execute a Task. This
will allow for other features of AME to be implemented while the
specification is refined.

Therefore this commit implements a simple but usable specification for a Task.