Implement Workflow spec reconciliation
Issue: 7

This commit implements support for the project-id and run-command
for in the Task specification. The controller will now generate
Workflows with appropriate arguments based on these two fields and
reconcile and correct workflows with Misconfigured arguments.

The test for Workflow argument reconciliation uses a sort of table
driven test with a wide range of cases meant to cover any possible sort
of misconfiguration. For the future it would probably be a good idea to
look into auto generating these cases, especially then the
specificationg grows more complex.

Also the table driven tests should be refacored when removing ginkgo to
follow the format suggested in the github wiki for go table driven tests
with go test. This will run each case as a separate test and make i
obvious which cases fails and why.