Implement gRPC service for AME tasks
Issue: 12

This commit implements a simple gRPC service which can create and get
Tasks aswell as code generation for protobuf definitions and k8s rest
clients based on the Task types.
This avoids a considerable amount of manual work every time the Task
types are updated. The CI has been updated to test and build container
images for the service.

Note:
that this is an initial minimally viable gRPC service and it should
be heavily refactored a long with the project structure as a whole in
future commits. The tooling around the protobuf and k8s REST client
autogeneration is hard to understand as a new comer and needs to be made
more approachable and maintainable.

The server/main.go file is just there to be able to compile and test
that the AME server , there is currently no way to configure server
externally this will be a task for a future commit.