Implement a minimal gRPC based CLI and e2e test
Issue: 14

In order to interact with AME a CLI is required.

Therefor this commit implements a minimal gRPC based CLI to get the
basic tooling setup for testing and building the CLI in an e2e manner.

Note that the implementation has been done as minimally as possible to
focus on getting the tooling setup properly and will therefore require
refactoring going forward as the CLI's capababilities are expanded.

A few issues were found with the controller when deployed to an actual
k8s cluster, these have been adressed as follows:
The argo workflow types were added to the controller's scheme and the
controller's service account permissions were corrected to allow for
monitoring and manipulating workflows in the cluster.