Implement server port configuration
Issue: 20

To make the server flexible it is important that it is configurable via
ennvironment variables.
Therefore this commit makes the port it listens on configurable via an
environment variable. A configmap and accompanying pod configuration has
been implemented aswell showing how to configure the server in a k8s
deployment.