Implement file storage in the server
Issue: 16

AME requires storage to be able to run Tasks for a project.
Therefore this commit implements support for preparing, storing,
downloading and clearing storage. The current implementation works with
the s3 API inorder to remain agnostic from the storage infrastructure is.
The commit also includes make file commands to start and stop a minio
instance used running in docker for testing the server storage.