Implement project uploads from the cli run command
Issue: 17

The AME CLI requires the ability to upload files before a task can be
executed. Therefore this commit implements that capability, the project
files are packaged in a single archive before sending. This supports
files being sent over a single stream and improves the
performance for large projects.

Sending a single large file also keeps the implementation somewhat
agnostic to how the project files are packaged, making it easier to
switch formats or add files compression during transport.