Implement token based authentication for gRPC
Issue: 21

In order to expose AME over a network authentication is a hard
requirement.
Therefore this commit implements support for setting a bearer token both in
the server and the client CLI. This allows for a simple authentication
mechanism until actual user management and oauth is implemented.

Note that the local cluster deployment for the server and the CLI tests
are both configured to set the same token in each environment for easy
testing, this is not meant as a long term solution but just an
intermediate setup to run the tests until a better setup is created for
configuring the local testing environment in a centralized manner.