Implement CLI configuration
Issue: 22

The CLI needs to be easily configurable for end users.

Therefore this commit implements support for configuration using
environment variables and a yaml config file in the user's home
directory. To avoid having to manually create the config file, this
commit also implements a setup command, guiding the user through the CLI
configuration.