Implement support for artifact sync and task logs
Issue: 19

Inorder for the task execution to be useful when running tasks from a
local machine with the CLI, artifacts produced during the run must be
detected and synchronized to the local machine and output from the
execution must be streamed to the local terminal, so the user knows
what is going on and can debug failures.

Therefore this commit implements support for downloading artifacts when
a task completes execution and streaming logs during execution.

Note that these are simple implementations that will be suffcient for
the prototype but must be refined.

The artifact synchronization to the local machine will have a hard time handling
large artifacts in the GB size range as the server loads all of the artifacts in
to memory at once. Also the artifact detection and storing relies
heavily on s3cmd and bash. This is very brittel, hard to maintain
and needs to be replaced with an go tool of some sort which can properly
tested and maintained.

The logging has some limitations around showing terminal animations properly
and does not currently store logs for tasks so they are gone once Argo
has garbarge collected the underlying Workflow and logs.

For future reference, this commit is way way too big