Implement support for a project configuration file
Issue: 36

In order to store project configuration the CLI needs to support
creating and editing an in project file.

Therefore this commit implements support for the CLI to generate an
ame.yaml file inside the project directory. As well as adding multiple
task specifications to that file.

Extras:

- A spinner has been implemented for the run command, indicating what is
happening.

- The memory limit has been increased for the server deployment, as it was
sometimes killed due to OOM. Once we have actual performance tests can we
investigate this properly.

- Request logging been implemented in the server.