Implement support for environment variable config
Issue: 35

Some tasks will be require the environment to be configured, by the user.
Therefore this commit implements support for configuring the environment
via the CLI and project file.

Note that in order to set the environment for the underlying Argo
Workflow, the workflow template has been replaced by the go code
generating the entire workflow, so the environment for the container can
be set.