Implement support for secrets
Issue: 35

Some tasks will require access to secrets, which can not be specified
in the project file for security reasons.
Therefore this commit implements support for referencing secrets and
having AME expose them to the task during execution.
To keep the initial implementation simple the secrets are k8s secrets
internally.