Implement support for Task piplines
Issue: 41

In order to chain multiple tasks together with different requirements
AME requires the ability to have a Task consist of mulitple sub tasks.

Therefore this commit implements support for pipelines, where a task can
specify a pipeline of multiple sub tasks which are executed
sequentially, while carriying over the projet state between tasks to
ensure that the project behaves as it would on a local machine.