Implement RecurringTask CRD and controller
Issue: 51

AME needs to support recurring tasks.

Therefore this commit implements a CRD and controller which facilitate
scheduling tasks to run based on a cron schedule.
This also resulted in some refactoring of the Task controller code and tests
to avoid carrying over bad practices from the original tests to the new
controller's test.

Extra:
- The Makefile has been refactoried to limit code generation.
- A recipe for testing the just the controllers has been included to speed up
  controller testing time.
- A recipe for reloading the server has been implemented, to help speed
  up testing time;