Implement Git project sources in the Task CRD
Issue: 54

In order for AME to have features like recurring Tasks, it is
important to support pulling in a project from a Git repository.

Therefore this commit implements support for specifiying a Git respository
and reference to use as a source for AME to fetch project files
and execute a Task. This feature has only been implemented in the
in the Task CRD and controller, it is not exposed through the CLI.
The gRPC server should support the new fields, but this has not been tested.

Extra:
Reworking logging was required to get this feature to work, as all Tasks
are now multistep under the hood, where the first step either clonse a
repo and checks out the reference, or it downloads the project from
object storage. Since logging was not working across multiple steps,
this broke a lot of tests. Therefore this commit also implements support
for logging Task piplines.