Prepare AME for the first demo
The demo preparation includes a number of changes to the CLI, improving
the experience of creating tasks, scheduling tasks and more.
Refactor create_task and add support for pipelines

The manifests have been updated to work with GKE, this require a few
things related to pull secrets, RBAC and container users.

The prepare_cluster.sh script has been included to easily reproduce a
single node GKE cluster with a GPU for demo purposes.