Get CLI tests working again
Issue: #78

This commit fixes failing CLI tests as preparation to reimplementing
the Rust controller in Rust.

The failures were caused by prepartion for latest demo.

A few tests were disabled, as they don't really carry over to the new
components

The changes are not polished as they will not be
longlived, due to the pending migration to rust.