Implement Task CRD and controller in Rust
Issue: #78

As a part of migrating AME to the new Rust based stack, the Task
controller and CRD must be reimplemented. Therefore this commit does
just that with a few omissions.
Some of the existing features have been left out, such as Task
pipelines, as these will change drastically due to recent feedback
and options with the new stack that weren't possible before.

Therefore a new Task specification which takes all this into account
will implemented in a separate commit.

This commit focuses on creating a solid foundation for building controllers
in Rust.

A few other tools have been setup in this commit aswell:
- justfile, this is a better alternative to Make for our usecase.
- common rust code quality tools.
- Updated CI to use these new tools.
- Kind has been replaced with K3S, as this allows for a more "real"
  testing experience.