Add common middlewares for gRPC service
Issue: #78

Before AME is put in production we must have logging, tracing and health
checks. Therefore this commit using Tower middleware to enable such
features.

Currently the health check is pretty barebones in that it just checks
that the server kan reach the Kubernetes API. This is currently the only
real requirement the server has for functioning. As the server grows in
compexity the health check should be expanded accordingly.