Implement new CLI
Issue: #78

This commits implements the Rust based CLI with a limited feature. The
features and polish are intentially limited here to focus on the new
project structure and tools. Polish and more features will be left for a
future commit.

The AME service was expanded to handling logging and project file
uploads inorder to faciliate the CLI features.