Refactor
This commit improves various parts of the new Rust based components.

A lot of different components are touched in this commit, normally that
would be unacceptable. However as we are sacaffolding a new project
structure this is hard to avoid.

There are too many changes to describe them all in detail but here are
the highlights:

- Get the Task logging to work across multiple pods.
- Add deployment manifests for minio.
- Correct deployment manifests for the controller.
- A new test project was created using the new AME file specification.