Integrate MLflow with Tasks
Issue: #82

In order to support projects using Mlflow running Task's must have
access to an MLflow tracking server.

Therefore this commit implements this integration as well as the option
to have AME execute the Mlflow project automtically.

Note that the mlflow deployment is not included in this commit and will
be committed separately. The helm chart used for testing is unlicensed
we therefore have to create our own. As there is an upcomming
deadline & demo we may postpone this until later.

This commit introduces a notion of Task types, allowing AME's controller
to determine if a Task should be executed by MLflow or not.

Extra:

Remove the action used to setup the just tool. It is now installed
directly with cargo. This removes an extra dependency in form of an
external Github action.