Implement Git repository tracking
issue: #92

AME must be able to track projects in a Git repository.

Therefore this commit implements a new controller and CRDs which facilitate
synchronizing project sources from a Git repositores. This includes support
for private repositories aswell. The tracking is currently limited to the
default branch but the architecture allows for this to easily be expanded
in a future commit.

Extra:

Add conditonal step to CI workflow for pushing a images with the latest
tag on the main branch.