Implement model deployment/inference with Mlflow
Issue: #93

Serving models over an API is an important requirement for AME.

Therefore this commit implements model serving using Mlflow's serving
capabilities. This includes watching Mlflow model registry to update
deployments to the latest model version as well as training a model if
no model is present.

Extra:

To facilitate testing this functionality a new RPC was exposed through
the CLI to easily add Git repositories as project sources.