Add Leptos web project
Issue: #97

AME requires a web based interface as well. Therefore this commit
implements a simple starter project using the Leptos framework.

As a POC the web project connects to AME's gRPC endpoint and makes a
request.

Extra:

The images are now built with nightly rust to accomodate Leptos. This
will be changes in a future commit, with some changes to the web project
that will remove the dependency on nightly rust.