Remove client crate
Issue: #121

To improve the code sharing and avoid circular dep
dependencies all library code is to be moved to
the library crate in /lib.

Therefore this commit moves the client crate's
functionality into /lib and refactors the rest
of the project as needed.

Extra:

Use cargo-binstall to speed up CI.